Move dispatcher into own module.
