+++ lib/sdk/model/core.js
@@ -10,57 +10,14 @@ module.metadata = {
 };
 
 const { Cu } = require("chrome");
-const method = require("method/core");
+const { dispatcher } = require("../util/dispatcher");
 const { curry, compose, partial, when,
-        method: asMethod, apply, constant } = require("../lang/functional");
+        method, apply, constant } = require("../lang/functional");
 const { viewFor } = require("../view/core");
 const { find } = require("../util/array");
 
 
-// Utility function that is just an enhancement over `method` to
-// allow predicate based dispatch in addition to polymorphic
-// dispatch. Unfortunately polymorphic dispatch does not quite
-// cuts it in the world of XPCOM where no types / classes exist
-// and all the XUL nodes share same type / prototype.
-// Probably this is more generic and belongs some place else, but
-// we can move it later once this will be relevant.
-let dispatcher = (hint) => {
-  let base = method(hint);
-  // Make a map for storing predicate, implementation mappings.
-  let implementations = new Map();
 
-  // Dispatcher function goes through `predicate, implementation`
-  // pairs to find predicate that matches first argument and
-  // returns application of arguments on the associated
-  // `implementation`. If no matching predicate is found delegates
-  // to a `base` polymorphic function.
-  let dispatch = (...args) => {
-    var value = args[0]
-    for (let [predicate, implementation] of implementations) {
-      if (predicate(value))
-        return implementation.apply(implementation, args);
-    }
-
-    return base.apply(base, args);
-  };
-
-  // TODO: Remove this line
-  dispatch.implementations = implementations;
-  // Expose base API.
-  dispatch.define = base.define;
-  dispatch.implement = base.implement;
-  dispatch.toString = base.toString;
-
-  // Add a `where` function to allow extending function via
-  // predicates.
-  dispatch.where = (predicate, implementation) => {
-    if (implementations.has(predicate))
-      throw TypeError("Already implemented for the given condition")
-    implementations.set(predicate, implementation);
-  }
-
-  return dispatch;
-};
 
 // Define `modelFor` accessor function that can be implemented
 // for different types of views. Since view's we'll be dealing
@@ -143,7 +100,7 @@ let warnDisposed = partial(console.warn,
 // attempts to get `view` associated with `this` pseudo variable, if
 // there is no `view` associated with it returns `null` otherwies invokes
 // `f` with `view` and rest of the arguments passed.
-exports.withView = f => asMethod(when(viewFor,
-                                      (model, ...params) =>
-                                        apply(f, viewFor(model), params),
+exports.withView = f => method(when(viewFor,
+                                    (model, ...params) =>
+                                    apply(f, viewFor(model), params),
                                       compose(constant(null), warnDisposed)));
